split the module multiagent-openmanus in the graph example
