+++ spring-ai-alibaba-graph-example/big-tool/src/main/java/com/alibaba/cloud/ai/graph/bigtool/BigToolApplication.java
@@ -1,0 +1,13 @@
+package com.alibaba.cloud.ai.graph.bigtool;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BigToolApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BigToolApplication.class, args);
+	}
+
+}
