+++ spring-ai-alibaba-graph-example/big-tool/src/main/resources/application.yml
@@ -1,0 +1,22 @@
+server:
+  port: 18080
+spring:
+  application:
+    name: big-tool
+  ai:
+    alibaba:
+      toolcalling:
+        weather:
+          enabled: true
+          api-key: ${WEATHER_API_KEY}
+    dashscope:
+      api-key: ${DASHSCOPE_API_KEY}
+    openai:
+      base-url: https://dashscope.aliyuncs.com/compatible-mode
+      api-key: ${DASHSCOPE_API_KEY}
+      chat:
+        options:
+          model: qwen-max-latest
+      embedding:
+        options:
+          model: text-embedding-v1
