+++ spring-ai-alibaba-graph-example/multiagent-openmanus/src/main/java/com/alibaba/cloud/ai/example/graph/GraphApplication.java
@@ -16,7 +16,7 @@
 
 package com.alibaba.cloud.ai.example.graph;
 
-import com.alibaba.cloud.ai.example.graph.react.tool.weather.function.WeatherProperties;
+import com.alibaba.cloud.ai.graph.react.function.WeatherProperties;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
