--- spring-ai-alibaba-graph-example/multiagent-openmanus/src/main/java/com/alibaba/cloud/ai/example/graph/bigtool/constants/Constant.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2024-2025 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.cloud.ai.example.graph.bigtool.constants;
-
-public class Constant {
-
-	public static final String INPUT_KEY = "input";
-
-	public static final String HIT_TOOL = "hit_tool";
-
-	public static final String SOLUTION = "solution";
-
-	public static final String TOOL_LIST = "tool_list";
-
-	public static final String METHOD_PARAMETER_TYPES = "parameter_types";
-
-	public static final String METHOD_NAME = "method_name";
-
-}
