--- spring-ai-alibaba-graph-example/multiagent-openmanus/src/main/java/com/alibaba/cloud/ai/example/graph/react/ReactAutoconfiguration.java
@@ -1,94 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.alibaba.cloud.ai.example.graph.react;
-
-import java.util.concurrent.TimeUnit;
-
-import com.alibaba.cloud.ai.graph.CompiledGraph;
-import com.alibaba.cloud.ai.graph.GraphRepresentation;
-import com.alibaba.cloud.ai.graph.exception.GraphStateException;
-import com.alibaba.cloud.ai.graph.agent.ReactAgent;
-import org.apache.hc.client5.http.classic.HttpClient;
-import org.apache.hc.client5.http.config.RequestConfig;
-import org.apache.hc.client5.http.impl.classic.HttpClients;
-import org.apache.hc.core5.util.Timeout;
-
-import org.springframework.ai.chat.client.ChatClient;
-import org.springframework.ai.chat.client.advisor.SimpleLoggerAdvisor;
-import org.springframework.ai.chat.model.ChatModel;
-import org.springframework.ai.openai.OpenAiChatOptions;
-import org.springframework.ai.tool.resolution.ToolCallbackResolver;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
-import org.springframework.web.client.RestClient;
-
-@Configuration
-public class ReactAutoconfiguration {
-
-	@Bean
-	public ReactAgent normalReactAgent(ChatModel chatModel, ToolCallbackResolver resolver) throws GraphStateException {
-		ChatClient chatClient = ChatClient.builder(chatModel)
-			.defaultToolNames("getWeatherFunction")
-			.defaultAdvisors(new SimpleLoggerAdvisor())
-			.defaultOptions(OpenAiChatOptions.builder().internalToolExecutionEnabled(false).build())
-			.build();
-
-		return ReactAgent.builder()
-			.name("React Agent Demo")
-			.chatClient(chatClient)
-			.resolver(resolver)
-			.maxIterations(10)
-			.build();
-	}
-
-	@Bean
-	public CompiledGraph reactAgentGraph(@Qualifier("normalReactAgent") ReactAgent reactAgent)
-			throws GraphStateException {
-
-		GraphRepresentation graphRepresentation = reactAgent.getStateGraph()
-			.getGraph(GraphRepresentation.Type.PLANTUML);
-
-		System.out.println("\n\n");
-		System.out.println(graphRepresentation.content());
-		System.out.println("\n\n");
-
-		return reactAgent.getAndCompileGraph();
-	}
-
-	@Bean
-	public RestClient.Builder createRestClient() {
-
-		// 2. 创建 RequestConfig 并设置超时
-		RequestConfig requestConfig = RequestConfig.custom()
-			.setConnectTimeout(Timeout.of(10, TimeUnit.MINUTES)) // 设置连接超时
-			.setResponseTimeout(Timeout.of(10, TimeUnit.MINUTES))
-			.setConnectionRequestTimeout(Timeout.of(10, TimeUnit.MINUTES))
-			.build();
-
-		// 3. 创建 CloseableHttpClient 并应用配置
-		HttpClient httpClient = HttpClients.custom().setDefaultRequestConfig(requestConfig).build();
-
-		// 4. 使用 HttpComponentsClientHttpRequestFactory 包装 HttpClient
-		HttpComponentsClientHttpRequestFactory requestFactory = new HttpComponentsClientHttpRequestFactory(httpClient);
-
-		// 5. 创建 RestClient 并设置请求工厂
-		return RestClient.builder().requestFactory(requestFactory);
-	}
-
-}
