+++ spring-ai-alibaba-graph-example/react/src/main/resources/application.yml
@@ -1,0 +1,23 @@
+server:
+  port: 8080
+spring:
+  application:
+    name: react
+  ai:
+    alibaba:
+      toolcalling:
+        weather:
+          enabled: true
+          api-key: ${WEATHER_API_KEY}
+    dashscope:
+      api-key: ${DASHSCOPE_API_KEY}
+    openai:
+      base-url: https://dashscope.aliyuncs.com/compatible-mode
+      api-key: ${DASHSCOPE_API_KEY}
+      chat:
+        options:
+          model: qwen-max-latest
+      embedding:
+        options:
+          model: text-embedding-v1
+
