+++ spring-ai-alibaba-graph-example/reflection/src/main/java/com/alibaba/cloud/ai/graph/reflection/ReflectionApplication.java
@@ -1,0 +1,13 @@
+package com.alibaba.cloud.ai.graph.reflection;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ReflectionApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ReflectionApplication.class, args);
+    }
+
+}
