+++ spring-ai-alibaba-graph-example/reflection/src/main/resources/application.yml
@@ -1,0 +1,17 @@
+server:
+  port: 8080
+spring:
+  application:
+    name: reflection
+  ai:
+    dashscope:
+      api-key: ${DASHSCOPE_API_KEY}
+    openai:
+      base-url: https://dashscope.aliyuncs.com/compatible-mode
+      api-key: ${DASHSCOPE_API_KEY}
+      chat:
+        options:
+          model: qwen-max-latest
+      embedding:
+        options:
+          model: text-embedding-v1
